Time: 11 ms (90.23%), Space: 8.8 MB (97.68%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -3,7 +3,7 @@ public:
     int numDistinct(string s, string t) {
         int m = s.size() ; 
         int n = t.size() ; 
-        vector<long long> dp(n+1 , 0);
+        vector<double> dp(n+1 , 0);
         dp[n] = 1 ; 
         for(int i=m-1 ; i>=0 ; i--){
             for(int j=0 ; j<=n ; j++){
